+++ operator/Makefile
@@ -44,17 +44,11 @@ licenses: ## Verifies dependency licenses and requires GITHUB_TOKEN to be set
 	go build $(GOFLAGS) -o bin/operator cmd/controller/main.go
 	golicense hack/license-config.hcl bin/operator
 
-apply: ## Deploy the controller into your ~/.kube/config cluster
-	kubectl create ns $(KIT_NAMESPACE) || true
-	kubectl -n $(KIT_NAMESPACE) apply -f config/
-
-deploy:
-	kubectl create ns $(KIT_NAMESPACE) || true
+deploy: ## Deploy the controller into your ~/.kube/config cluster
 	KO_DOCKER_REPO=$(CONTAINER_IMAGE_REGISTRY)/kit ko apply --bare -f config/
     
 delete: ## Delete the controller from your ~/.kube/config cluster
 	kubectl delete -f config
-	kubectl delete ns $(KIT_NAMESPACE)
 
 codegen: ## Generate code. Must be run if changes are made to ./pkg/apis/...
 	hack/codegen.sh
